test(semaphore): keep node reclamation checks next to state
